+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.12.1]
+### Added
+- SevenNet-Omni-i8, SevenNet-Omni-i12
+- L2MAE loss
+- OrderedSampler, batch training
+
 ### Fixed
 - FlashTP with LAMMPS parallel in torch
 
@@ -10,8 +15,6 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Documentation moved to RTD
 - LAMMPS-MLIAP integration with GhostExchangeOp
-
-### Added
 - SevenNet-Omni
 - Example config for fine-tuning the SevenNet-MF-ompa model
 - FlashTP support (https://github.com/SNU-ARC/flashTP)